# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Garfex API
  version: 0.1.0
  description: Public operational, API-reference, catalog-descriptor, catalog-list, and catalog-detail endpoints.
paths:
...
  /v1/catalog/descriptors:
    get:
      summary: List catalog descriptors
      description: Returns Core's dynamic catalog metadata. It does not list catalog records.
      responses:
        '200':
          description: Catalog descriptors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogDescriptorsResponse'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/Internal' }
  /v1/catalog/classes:
    get:
      summary: List active classes
      description: Returns every active CLASE record, unpaginated.
      responses:
        '200':
          description: Active CLASE records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveClassesResponse'
        '500': { $ref: '#/components/responses/Internal' }
...
  /v1/suppliers:
    get:
      summary: Search suppliers
      description: Returns a page of suppliers matching the query.
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE, ALL], default: ACTIVE }
        - name: text
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Supplier page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierPage'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '500': { $ref: '#/components/responses/Internal' }
    post:
      summary: Create a supplier
      description: Creates one supplier. Created suppliers are always active.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreateRequest'
      responses:
        '201':
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/Internal' }
  /v1/suppliers/from-cfdi/preview:
    post:
      summary: Preview creating a supplier from a CFDI XML
      description: >-
        Reads a CFDI 4.0 voucher (multipart `file` field or raw XML body) and
        returns the supplier draft built from its Emisor plus the supplier that
        already owns that RFC, active or inactive, or null when the RFC is new.
        Read-only: nothing is created. To create, call `POST /v1/suppliers`
        with the draft; the unique RFC constraint stays authoritative if
        another request creates the supplier in between.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The CFDI XML file.
          application/xml:
            schema:
              type: string
              description: The raw CFDI XML document.
      responses:
        '200':
          description: Draft and existing supplier, if any
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierCFDIPreview'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '422':
          description: >-
            The file is not a usable CFDI 4.0. `code` is one of INVALID_XML,
            NOT_CFDI, UNSUPPORTED_VERSION, INVALID_CFDI.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '500': { $ref: '#/components/responses/Internal' }
...
  /v1/resources:
    get:
      summary: Search resources
      description: Returns a page of resources matching the query.
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE, ALL], default: ACTIVE }
        - name: text
          in: query
          schema: { type: string }
        - name: classCode
          in: query
          schema: { type: string }
        - name: familyCode
          in: query
          schema: { type: string }
        - name: typeCode
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Resource page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePage'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '500': { $ref: '#/components/responses/Internal' }
    post:
      summary: Create a resource
      description: Creates one resource under a class/family/type scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Created resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/Internal' }
...
  /v1/cfdi/parse:
    post:
      summary: Parse a CFDI 4.0 XML
      description: >-
        Reads one CFDI 4.0 voucher (Mexican electronic invoice) and returns all
        of its data. Nothing is stored. The XML may be uploaded as a multipart
        `file` field or sent as the raw request body. `supplierDraft` holds
        the Emisor shaped like a supplier create request, so a client can
        prefill a supplier form; `taxRegime` has no supplier field and is
        informational.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The CFDI XML file.
          application/xml:
            schema:
              type: string
              description: The raw CFDI XML document.
          text/xml:
            schema:
              type: string
              description: The raw CFDI XML document.
      responses:
        '200':
          description: Parsed voucher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CFDIParseResponse'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '422':
          description: >-
            The file is not a usable CFDI 4.0. `code` is one of INVALID_XML,
            NOT_CFDI, UNSUPPORTED_VERSION, INVALID_CFDI.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '500': { $ref: '#/components/responses/Internal' }
  /v1/purchases:
    post:
      summary: Import a CFDI purchase document
      description: >-
        Imports one CFDI 4.0 purchase document from its raw XML, sent as
        multipart/form-data. Idempotent on the document's fiscal UUID:
        re-importing the same document returns 200 with alreadyExisted true
        and creates nothing; importing the same UUID with different content
        returns 409. If the Emisor RFC matches no existing supplier, a
        minimal supplier is auto-registered.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, actor]
              properties:
                file: { type: string, format: binary, description: The CFDI XML file. }
                actor: { type: string }
                branchId: { type: string, description: 'Optional strict positive decimal branch identifier.' }
      responses:
        '201':
          description: Purchase imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseImportResponse'
        '200':
          description: Purchase already existed with matching content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseImportResponse'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
...
components:
  responses:
    InvalidArgument: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Requested resource was not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Request failed validation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unavailable: { description: Service is temporarily unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Internal: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    CatalogRecord:
      type: object
      required: [kind, id, revision, active, values, rules]
      properties:
        kind: { type: string }
        id: { type: string, description: Decimal integer identifier, pattern: '^(?:0|-?[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        revision: { type: string, description: Decimal unsigned revision, pattern: '^(?:0|[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        active: { type: boolean }
        values: { type: object, additionalProperties: { $ref: '#/components/schemas/CatalogValue' } }
        rules: { type: array, items: { $ref: '#/components/schemas/ApplicabilityRule' } }
...
    CatalogDescriptorsResponse:
      type: object
      required: [descriptors]
      properties: { descriptors: { type: array, items: { $ref: '#/components/schemas/CatalogDescriptor' } } }
    ActiveClassesResponse:
      type: object
      required: [records]
      properties: { records: { type: array, items: { $ref: '#/components/schemas/CatalogRecord' } } }
    CatalogDescriptor:
      type: object
      required: [kind, singular, plural, fields, identityFields, parentKind, parentField]
      properties:
        kind: { type: string }
        singular: { type: string }
        plural: { type: string }
        fields: { type: array, items: { $ref: '#/components/schemas/FieldDescriptor' } }
        identityFields: { type: array, items: { type: string } }
        parentKind: { type: string }
        parentField: { type: string }
...
    SupplierCFDIPreview:
      type: object
      additionalProperties: false
      required: [draft, existing]
      properties:
        draft: { $ref: '#/components/schemas/CFDISupplierDraft' }
        existing:
          nullable: true
          description: The supplier already owning the RFC (active or inactive); null when the RFC is new.
          allOf:
            - $ref: '#/components/schemas/Supplier'
    CFDIParseResponse:
      type: object
      additionalProperties: false
      required: [invoice, supplierDraft]
      properties:
        invoice: { $ref: '#/components/schemas/CFDIInvoice' }
        supplierDraft: { $ref: '#/components/schemas/CFDISupplierDraft' }
    CFDISupplierDraft:
      type: object
      additionalProperties: false
      required: [taxIdentifier, legalName, taxRegime]
      description: Emisor data shaped like a supplier create request. RFC is trimmed and upper-cased.
      properties:
        taxIdentifier: { type: string }
        legalName: { type: string }
        taxRegime: { type: string, description: "SAT regimen fiscal code, such as 601. Informational only; suppliers have no such field." }
    CFDIInvoice:
      type: object
      additionalProperties: false
      description: >-
        Amounts, quantities and rates are exact decimal strings as written in
        the XML. Absent values are empty strings. Dates have no zone
        (`YYYY-MM-DDThh:mm:ss`).
      required:
        - version
        - series
        - folio
        - issuedAt
        - seal
        - paymentForm
        - certificateNumber
        - certificate
        - paymentConditions
        - subtotal
        - discount
        - currency
        - exchangeRate
        - total
        - voucherType
        - export
        - paymentMethod
        - issuePlace
        - confirmation
        - related
        - issuer
        - receiver
        - concepts
        - taxes
        - stamp
      properties:
        version: { type: string }
        series: { type: string }
        folio: { type: string }
        issuedAt: { type: string }
        seal: { type: string }
        paymentForm: { type: string }
        certificateNumber: { type: string }
        certificate: { type: string }
        paymentConditions: { type: string }
        subtotal: { type: string }
        discount: { type: string }
        currency: { type: string }
        exchangeRate: { type: string }
        total: { type: string }
        voucherType: { type: string }
        export: { type: string }
        paymentMethod: { type: string }
        issuePlace: { type: string }
        confirmation: { type: string }
        related:
          type: array
          items: { $ref: '#/components/schemas/CFDIRelated' }
        issuer: { $ref: '#/components/schemas/CFDIIssuer' }
        receiver: { $ref: '#/components/schemas/CFDIReceiver' }
        concepts:
          type: array
          items: { $ref: '#/components/schemas/CFDIConcept' }
        taxes: { $ref: '#/components/schemas/CFDITaxes' }
        stamp:
          nullable: true
          description: The SAT TimbreFiscalDigital complement; null when the voucher has none.
          allOf:
            - $ref: '#/components/schemas/CFDIStamp'
    CFDIRelated:
      type: object
      additionalProperties: false
      required: [relationType, uuids]
      properties:
        relationType: { type: string }
        uuids:
          type: array
          items: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 'Generic, sanitized classification of the failure (e.g. "invalid request", "validation failed").'
        code:
          type: string
          description: >-
            Stable, machine-readable reason code, safe to switch on: distinct
            outcomes that read the same generic "conflict" error (e.g. an
            expected-revision mismatch, a still-active record, or a record
            still in use) are always distinguishable here even when error is
            identical. Absent only when the failure never reached Core (e.g.
            an HTTP method mismatch).
          enum:
            - INVALID_ARGUMENT
            - NOT_FOUND
            - DUPLICATE
            - INVALID_REFERENCE
            - VALIDATION
            - INTEGRITY
            - IDENTITY_CONFLICT
            - INVALID_LIFECYCLE
            - REACTIVATION_IMPOSSIBLE
            - INVALID_CATALOG
            - IN_USE
            - IMMUTABLE_CODE
            - CONFLICT
            - UNAVAILABLE
            - INTERNAL
        detail:
          type: string
          description: >-
            The specific reason, present only when error is "invalid request" or
            "validation failed". Absent for every other error (not found, conflict,
            unavailable, internal server error), which never expose internal detail.
    SupplierCreateRequest:
      type: object
      additionalProperties: false
      required: [actor]
      properties:
        actor: { type: string }
        tradeName: { type: string }
        legalName: { type: string }
        taxIdentifier: { type: string }
        website: { type: string }
        notes: { type: string }
...
    Supplier:
      type: object
      additionalProperties: false
      required: [id, tradeName, legalName, taxIdentifier, website, notes, active, createdAt, updatedAt]
      properties:
        id: { type: string, description: Decimal integer identifier, pattern: '^(?:0|-?[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        tradeName: { type: string }
        legalName: { type: string }
        taxIdentifier: { type: string }
        website: { type: string }
        notes: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SupplierPage:
      type: object
      additionalProperties: false
      required: [suppliers, hasPrevious, hasNext]
      properties:
        suppliers: { type: array, items: { $ref: '#/components/schemas/Supplier' } }
        hasPrevious: { type: boolean }
        hasNext: { type: boolean }
...
    ResourceScope:
      type: object
      additionalProperties: false
      required: [classCode, familyCode, typeCode]
      properties:
        classCode: { type: string }
        familyCode: { type: string }
        typeCode: { type: string }
    ResourceAttribute:
      type: object
      additionalProperties: false
      required: [code, value]
      properties:
        code: { type: string }
        value: { $ref: '#/components/schemas/CatalogValue' }
    ResourceCreateRequest:
      type: object
      additionalProperties: false
      required: [actor, scope, naturalUnit, attributes]
      properties:
        actor: { type: string }
        scope: { $ref: '#/components/schemas/ResourceScope' }
        naturalUnit: { type: string }
        attributes: { type: array, items: { $ref: '#/components/schemas/ResourceAttribute' } }
...
    Resource:
      type: object
      additionalProperties: false
      required: [id, identityV1, scope, naturalUnit, active, revision, attributes]
      properties:
        id: { type: string, description: Decimal integer identifier, pattern: '^(?:0|-?[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        identityV1: { type: string }
        scope: { $ref: '#/components/schemas/ResourceScope' }
        naturalUnit: { type: string }
        active: { type: boolean }
        revision: { type: string, description: Decimal unsigned revision, pattern: '^(?:0|[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        attributes: { type: array, items: { $ref: '#/components/schemas/ResourceAttribute' } }
    ResourcePage:
      type: object
      additionalProperties: false
      required: [resources, hasPrevious, hasNext]
      properties:
        resources: { type: array, items: { $ref: '#/components/schemas/Resource' } }
        hasPrevious: { type: boolean }
        hasNext: { type: boolean }
...
    PurchaseImportResponse:
      type: object
      additionalProperties: false
      required: [purchase, lines, alreadyExisted]
      properties:
        purchase: { $ref: '#/components/schemas/Purchase' }
        lines: { type: array, items: { $ref: '#/components/schemas/PurchaseLine' } }
        alreadyExisted: { type: boolean, description: True when the document's fiscal UUID was already registered with matching content; no additional write happened. }
...